Custo e Força adicionado aos botões
Dentro do script de clique, mudei as variáveis para float.
Criei uma classe para gerenciar a força do clique e adicionar o custo aos botões.

+++ Clicker-modelagem/Assets/clique.cs
@@ -6,8 +6,8 @@ using UnityEngine.UI;
 public class clique : MonoBehaviour {
 
     public Text score;
-    public int Nresources =0;
-    public int resourcesPerClick = 1;
+    public float Nresources =0.00f;
+    public float resourcesPerClick = 1;
